+++ src/modules/accounts/infra/typeorm/repositories/UsersRepository.ts
@@ -1,7 +1,9 @@
 import { getRepository, Repository } from 'typeorm';
-import { ICreateUserDTO } from '../../dtos/ICreateUserDTO';
-import { User } from '../../entities/User';
-import { IUsersRepository } from '../IUsersRepository';
+
+import { User } from '../entities/User';
+import { ICreateUserDTO } from '@modules/accounts/dtos/ICreateUserDTO';
+import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+
 
 class UsersRepository implements IUsersRepository {
   private repository: Repository<User>;
