+++ src/modules/accounts/repositories/in-memory/UsersRepositoryInMemory.ts
@@ -1,6 +1,7 @@
-import { ICreateUserDTO } from '../../dtos/ICreateUserDTO';
-import { User } from '../../entities/User';
-import { IUsersRepository } from '../IUsersRepository';
+import { ICreateUserDTO } from "@modules/accounts/dtos/ICreateUserDTO";
+import { User } from "@modules/accounts/infra/typeorm/entities/User";
+import { IUsersRepository } from "../IUsersRepository";
+
 
 class UsersRepositoryInMemory implements IUsersRepository {
   users: User[] = [];
