+++ src/modules/cars/infra/typeorm/entities/CategoryModel.ts
