+++ src/modules/cars/infra/typeorm/repositories/SpecificationsRepository.ts
@@ -3,11 +3,12 @@
 /* eslint-disable import/prefer-default-export */
 
 import { getRepository, Repository } from 'typeorm';
-import { Specification } from '../../entities/Specification';
+
+import { Specification } from '../entities/Specification';
 import {
-  ISpecificationsRepository,
   ICreateSpecificationsDTO,
-} from '../ISpecificationsRepository';
+  ISpecificationsRepository,
+} from '@modules/cars/repositories/ISpecificationsRepository';
 
 class SpecificationsRepository implements ISpecificationsRepository {
   private repository: Repository<Specification>;
