+++ src/modules/cars/useCases/importCategory/ImportCategoryUseCase.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import csvParse from 'csv-parse';
 import { inject, injectable } from 'tsyringe';
 
-import { CategoriesRepository } from '@modules/cars/repositories/implementations/CategoriesRepository';
+import { CategoriesRepository } from '@modules/cars/infra/typeorm/repositories/CategoriesRepository';
 
 interface IImportCategory {
   name: string;
