+++ src/shared/infra/http/routes/categories.routes.ts
