+++ src/shared/infra/http/server.ts
@@ -2,14 +2,14 @@ import express, { NextFunction, Request, Response } from 'express';
 import 'express-async-errors';
 import swaggerUi from 'swagger-ui-express';
 
-import './database';
-
-import './shared/container';
+import '../../../database';
 
 import { router } from './routes';
 
-import swaggerFile from './swagger.json';
-import { AppError } from './errors/AppError';
+import '../../container';
+import { AppError } from '@errors/AppError';
+
+import swaggerFile from '../../../swagger.json';
 
 const app = express();
 const port = 3000;
